backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
boto3==1.40.7
llama-index==0.13.1
llama-index-retrievers-bm25==0.6.3
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
PyMuPDF==1.26.3
pydantic-settings==2.10.1
